requirements.txt
functions-framework>=3.0
slack-bolt>=1.21.0
google-genai>=1.0.0
PyGithub>=2.1.0
